Running savefile

+++ Data/EasySave/JsonObjects/RunningLogJson.cs
@@ -11,7 +11,7 @@ namespace EasySave
         public string TargetFilePath { get; set; }
         public string State { get; set; }
         public int TotalFilesToCopy { get; set; }
-        public int TotalFilesSize { get; set; }
+        public long TotalFilesSize { get; set; }
         public int NbFilesLeftToDo { get; set; }
         public int Progression { get; set; }
